refactor(components): 'CategoryFilterMenu'에서 'CategoryEditModal' 분리
